# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  provider: "provider",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterBodyRole;
  phone?: string;
  vehicleType?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  provider: "provider",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  vehicleType?: string | null;
  avatarUrl?: string | null;
  rating?: number | null;
  totalRatings: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateRequestBodyServiceType =
  (typeof CreateRequestBodyServiceType)[keyof typeof CreateRequestBodyServiceType];

export const CreateRequestBodyServiceType = {
  delivery: "delivery",
  transport: "transport",
  logistics: "logistics",
} as const;

export interface CreateRequestBody {
  pickupLocation: string;
  dropLocation: string;
  pickupLat?: number | null;
  pickupLng?: number | null;
  dropLat?: number | null;
  dropLng?: number | null;
  description: string;
  serviceType: CreateRequestBodyServiceType;
  offeredPrice?: number | null;
  distanceKm?: number | null;
  scheduledAt?: string | null;
}

export type EstimateBodyServiceType =
  (typeof EstimateBodyServiceType)[keyof typeof EstimateBodyServiceType];

export const EstimateBodyServiceType = {
  delivery: "delivery",
  transport: "transport",
  logistics: "logistics",
} as const;

export interface EstimateBody {
  pickupLocation: string;
  dropLocation: string;
  distanceKm: number;
  serviceType: EstimateBodyServiceType;
}

export type PriceEstimateBreakdownItem = {
  label: string;
  amount: number;
};

export interface PriceEstimate {
  baseFare: number;
  distanceCharge: number;
  serviceFee: number;
  total: number;
  breakdown: PriceEstimateBreakdownItem[];
  distanceKm: number;
}

export type ServiceRequestServiceType =
  (typeof ServiceRequestServiceType)[keyof typeof ServiceRequestServiceType];

export const ServiceRequestServiceType = {
  delivery: "delivery",
  transport: "transport",
  logistics: "logistics",
} as const;

export type ServiceRequestStatus =
  (typeof ServiceRequestStatus)[keyof typeof ServiceRequestStatus];

export const ServiceRequestStatus = {
  requested: "requested",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface ServiceRequest {
  id: number;
  customerId: number;
  providerId?: number | null;
  pickupLocation: string;
  dropLocation: string;
  pickupLat?: number | null;
  pickupLng?: number | null;
  dropLat?: number | null;
  dropLng?: number | null;
  description: string;
  serviceType: ServiceRequestServiceType;
  status: ServiceRequestStatus;
  offeredPrice?: number | null;
  agreedPrice?: number | null;
  distanceKm?: number | null;
  scheduledAt?: string | null;
  completedAt?: string | null;
  customer: User;
  provider?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface RequestList {
  requests: ServiceRequest[];
  total: number;
}

export type UpdateStatusBodyStatus =
  (typeof UpdateStatusBodyStatus)[keyof typeof UpdateStatusBodyStatus];

export const UpdateStatusBodyStatus = {
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}

export interface Agreement {
  id: number;
  requestId: number;
  terms: string;
  agreedPrice: number;
  customerSigned: boolean;
  providerSigned: boolean;
  customerSignedAt?: string | null;
  providerSignedAt?: string | null;
  fullyExecuted: boolean;
  createdAt: string;
}

export interface AgreementList {
  agreements: Agreement[];
}

export interface TrackingUpdate {
  id: number;
  requestId: number;
  status: string;
  message: string;
  lat?: number | null;
  lng?: number | null;
  createdAt: string;
}

export interface TrackingData {
  requestId: number;
  currentStatus: string;
  updates: TrackingUpdate[];
  currentLat?: number | null;
  currentLng?: number | null;
}

export interface TrackingUpdateBody {
  status: string;
  message: string;
  lat?: number | null;
  lng?: number | null;
}

export interface Rating {
  id: number;
  requestId: number;
  raterId: number;
  ratedUserId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  review?: string | null;
  createdAt: string;
}

export interface RatingList {
  ratings: Rating[];
  averageScore: number;
  totalCount: number;
}

export interface CreateRatingBody {
  requestId: number;
  ratedUserId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  review?: string | null;
}

export type DisputeStatus = (typeof DisputeStatus)[keyof typeof DisputeStatus];

export const DisputeStatus = {
  open: "open",
  under_review: "under_review",
  resolved: "resolved",
  closed: "closed",
} as const;

export interface Dispute {
  id: number;
  requestId: number;
  raisedById: number;
  reason: string;
  description: string;
  status: DisputeStatus;
  resolution?: string | null;
  createdAt: string;
  resolvedAt?: string | null;
  request?: ServiceRequest | null;
}

export interface DisputeList {
  disputes: Dispute[];
}

export interface CreateDisputeBody {
  requestId: number;
  reason: string;
  description: string;
}

export type ResolveDisputeBodyStatus =
  (typeof ResolveDisputeBodyStatus)[keyof typeof ResolveDisputeBodyStatus];

export const ResolveDisputeBodyStatus = {
  resolved: "resolved",
  closed: "closed",
} as const;

export interface ResolveDisputeBody {
  resolution: string;
  status: ResolveDisputeBodyStatus;
}

export interface CustomerDashboard {
  totalRequests: number;
  activeRequests: number;
  completedRequests: number;
  totalSpent: number;
  recentRequests: ServiceRequest[];
  openDisputes: number;
}

export interface ProviderDashboard {
  totalDeliveries: number;
  activeDeliveries: number;
  completedDeliveries: number;
  totalEarnings: number;
  averageRating: number;
  recentDeliveries: ServiceRequest[];
  openDisputes: number;
  pendingRequests: number;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  requestId?: number | null;
  createdAt: string;
}

export interface NotificationList {
  notifications: Notification[];
  unreadCount: number;
}

export type ListRequestsParams = {
  status?: ListRequestsStatus;
  role?: ListRequestsRole;
  limit?: number;
  offset?: number;
};

export type ListRequestsStatus =
  (typeof ListRequestsStatus)[keyof typeof ListRequestsStatus];

export const ListRequestsStatus = {
  requested: "requested",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListRequestsRole =
  (typeof ListRequestsRole)[keyof typeof ListRequestsRole];

export const ListRequestsRole = {
  customer: "customer",
  provider: "provider",
} as const;