# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChainTrack Supply Chain Transparency Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /agreements:
    get:
      operationId: listAgreements
      tags: [agreements]
      summary: List agreements
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of agreements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgreementList"
  /agreements/{requestId}:
    get:
      operationId: getAgreement
      tags: [agreements]
      summary: Get agreement for a request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agreement detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
  /agreements/{requestId}/sign:
    post:
      operationId: signAgreement
      tags: [agreements]
      summary: Sign a digital agreement
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
...
  /ratings:
    post:
      operationId: createRating
      tags: [ratings]
      summary: Submit a rating
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingBody"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
  /ratings/user/{userId}:
    get:
      operationId: getUserRatings
      tags: [ratings]
      summary: Get ratings for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingList"
  /disputes:
    get:
      operationId: listDisputes
      tags: [disputes]
      summary: List disputes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Disputes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeList"

    post:
      operationId: createDispute
      tags: [disputes]
      summary: Raise a dispute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDisputeBody"
      responses:
        "201":
          description: Dispute raised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{id}:
    get:
      operationId: getDispute
      tags: [disputes]
      summary: Get dispute by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispute detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /disputes/{id}/resolve:
    patch:
      operationId: resolveDispute
      tags: [disputes]
      summary: Resolve a dispute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, provider]
        phone:
          type: string
        vehicleType:
          type: string
          nullable: true
      required: [name, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, provider]
        phone:
          type: string
          nullable: true
        vehicleType:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalRatings:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, totalRatings, createdAt]
...
    ServiceRequest:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        providerId:
          type: integer
          nullable: true
        pickupLocation:
          type: string
        dropLocation:
          type: string
        pickupLat:
          type: number
          nullable: true
        pickupLng:
          type: number
          nullable: true
        dropLat:
          type: number
          nullable: true
        dropLng:
          type: number
          nullable: true
        description:
          type: string
        serviceType:
          type: string
          enum: [delivery, transport, logistics]
        status:
          type: string
          enum: [requested, accepted, in_progress, completed, cancelled]
        offeredPrice:
          type: number
          nullable: true
        agreedPrice:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        customer:
          $ref: "#/components/schemas/User"
        provider:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerId, pickupLocation, dropLocation, description, serviceType, status, createdAt, updatedAt, customer]
...
    Agreement:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        terms:
          type: string
        agreedPrice:
          type: number
        customerSigned:
          type: boolean
        providerSigned:
          type: boolean
        customerSignedAt:
          type: string
          format: date-time
          nullable: true
        providerSignedAt:
          type: string
          format: date-time
          nullable: true
        fullyExecuted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, requestId, terms, agreedPrice, customerSigned, providerSigned, fullyExecuted, createdAt]
    AgreementList:
      type: object
      properties:
        agreements:
          type: array
          items:
            $ref: "#/components/schemas/Agreement"
      required: [agreements]
...
    Rating:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        raterId:
          type: integer
        ratedUserId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, requestId, raterId, ratedUserId, score, createdAt]
    RatingList:
      type: object
      properties:
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/Rating"
        averageScore:
          type: number
        totalCount:
          type: integer
      required: [ratings, averageScore, totalCount]
    CreateRatingBody:
      type: object
      properties:
        requestId:
          type: integer
        ratedUserId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
          nullable: true
      required: [requestId, ratedUserId, score]
    Dispute:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        raisedById:
          type: integer
        reason:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, resolved, closed]
        resolution:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        request:
          $ref: "#/components/schemas/ServiceRequest"
          nullable: true
      required: [id, requestId, raisedById, reason, description, status, createdAt]
    DisputeList:
      type: object
      properties:
        disputes:
          type: array
          items:
            $ref: "#/components/schemas/Dispute"
      required: [disputes]
    CreateDisputeBody:
      type: object
      properties:
        requestId:
          type: integer
        reason:
          type: string
        description:
          type: string
      required: [requestId, reason, description]
    ResolveDisputeBody:
      type: object
      properties:
        resolution:
          type: string
        status:
          type: string
          enum: [resolved, closed]
      required: [resolution, status]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        requestId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, message, isRead, createdAt]
    NotificationList:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required: [notifications, unreadCount]