# lib/api-zod/src/generated/types/agreement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Agreement {
  id: number;
  requestId: number;
  terms: string;
  agreedPrice: number;
  customerSigned: boolean;
  providerSigned: boolean;
  customerSignedAt?: Date | null;
  providerSignedAt?: Date | null;
  fullyExecuted: boolean;
  createdAt: Date;
}