# lib/api-zod/src/generated/types/createRatingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRatingBody {
  requestId: number;
  ratedUserId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  review?: string | null;
}