# lib/api-zod/src/generated/types/ratingList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Rating } from "./rating";

export interface RatingList {
  ratings: Rating[];
  averageScore: number;
  totalCount: number;
}