# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainTrack Supply Chain Transparency Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  vehicleType?: string | null;
  avatarUrl?: string | null;
  rating?: number | null;
  totalRatings: number;
  createdAt: Date;
}